+++ tests/unit/__init__.py
@@ -7,6 +7,7 @@ import asyncprawcore
 
 class UnitTest(asynctest.TestCase):
     """Base class for asyncprawcore unit tests."""
+
     async def setUp(self) -> None:
         self.requestor = asyncprawcore.requestor.Requestor(
             "asyncprawcore:test (by /u/Lil_SpazJoekp)"